+++ crates/agglayer-interop-types/src/aggchain_proof/test.rs
@@ -1,0 +1,18 @@
+use super::*;
+
+#[test]
+fn regression_sp1_serialization_roundtrip_fail() {
+    // Conclusion: sp1 serialization is not deterministic, removed the equality
+    // check.
+    use bincode::Options;
+    let bytes = hex::decode("00000000000000000000000000000000000000000000fb00000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000").unwrap();
+    let input: SP1StarkProof = bincode::options()
+        .deserialize(&bytes)
+        .expect("failed first deserialization, would be fine");
+    let serialized: Vec<u8> = crate::bincode::default()
+        .serialize(&input)
+        .expect("failed serialization, unexpected");
+    let _output: SP1StarkProof = crate::bincode::default()
+        .deserialize(&serialized)
+        .expect("failed second deserialization, is unexpected");
+}
